+++ src/utils.c
@@ -1,6 +1,6 @@
 #include <utils.h>
 
-int check_memory_allocation(void *allocated_data){
+int8_t check_memory_allocation(void *allocated_data){
     if(allocated_data == NULL) {
         fprintf(stderr, "Erreur d'allocation mémoire\n");
         return EXIT_FAILURE;
